DEVHUB-485: Create a Generic Image Dropzone Component (#687)
* Move main image dropzone to shared component

* Use general image dropzone for both

* Simplify change handlers

* Remove source comments

* Prevent call stack exceeded
